Docs: Documentation all code.

+++ solution/microservices/ad_engine/src/domain/services/redis_service.rs
@@ -2,17 +2,20 @@ use serde::Serialize;
 
 use crate::{domain, infrastructure};
 
+/// Redis service for managing campaign data and moderation settings
 pub struct RedisService<'p> {
     repo: infrastructure::cash::redis::RedisExecutor<'p>,
 }
 
 impl<'p> RedisService<'p> {
+    /// Creates a new RedisService instance
     pub fn new(pool: &'p infrastructure::database_connection::redis::RedisPool) -> Self {
         RedisService {
             repo: infrastructure::cash::redis::RedisExecutor::new(pool),
         }
     }
 
+    /// Retrieves all active campaigns from Redis
     pub async fn get_all_active_campaigns(
         &self,
     ) -> domain::services::ServiceResult<Vec<domain::schemas::ActiveCampaignSchema>> {
@@ -50,6 +53,8 @@ impl<'p> RedisService<'p> {
         Ok(active_campaigns)
     }
 
+    /// Extracts a 24 character random chunk from a UUID
+    #[inline]
     async fn get_random_chunk_from_uuid(&self, id: &uuid::Uuid) -> String {
         id.to_string()
             .replace("-", "")
@@ -59,11 +64,13 @@ impl<'p> RedisService<'p> {
             .collect::<String>()
     }
 
+    /// Deletes an active campaign by ID
     pub async fn del_active_campaigns(&self, id: &uuid::Uuid) -> domain::services::ServiceResult<()> {
         let random_id = self.get_random_chunk_from_uuid(id).await;
         self.repo.delete(&format!("active_campaign:{random_id}")).await
     }
 
+    /// Retrieves a single active campaign by ID
     pub async fn get_active_campaign(
         &self,
         id: &uuid::Uuid,
@@ -72,6 +79,7 @@ impl<'p> RedisService<'p> {
         self.repo.get(&format!("active_campaign:{random_id}")).await
     }
 
+    /// Stores an active campaign
     pub async fn set_active_campaign(
         &self,
         data: domain::schemas::ActiveCampaignSchema,
@@ -80,29 +88,35 @@ impl<'p> RedisService<'p> {
         self.repo.set(&format!("active_campaign:{random_id}"), data).await
     }
 
+    /// Retrieves the list of obscene words used for moderation
     pub async fn get_obscene_words(&self) -> domain::services::ServiceResult<Vec<String>> {
         let word: String = self.repo.get("obscene_words").await?;
         Ok(word.split(',').map(|s| s.to_string()).collect())
     }
 
+    /// Updates the list of obscene words used for moderation
     pub async fn set_obscene_words(&self, data: Vec<String>) -> domain::services::ServiceResult<()> {
         self.repo.set("obscene_words", data.join(",").to_string()).await
     }
 
+    /// Gets the advance time setting, defaulting to 0 if not set
     pub async fn get_advance_time(&self) -> domain::services::ServiceResult<u32> {
         match self.repo.get("advance_time").await {
             Ok(data) => Ok(data),
             Err(_) => {
-                self.set_advance_time(0 as u32).await?;
+                self.set_advance_time(0).await?;
                 self.repo.get("advance_time").await
             },
         }
     }
 
+    /// Sets the advance time setting
     pub async fn set_advance_time(&self, data: u32) -> domain::services::ServiceResult<()> {
         self.repo.set("advance_time", data).await
     }
 
+    /// Gets the auto-moderation activation status, defaulting to false if not
+    /// set
     pub async fn get_is_activate_auto_moderate(&self) -> domain::services::ServiceResult<bool> {
         match self.repo.get("is_activate_auto_moderate").await {
             Ok(data) => Ok(data),
@@ -113,6 +127,7 @@ impl<'p> RedisService<'p> {
         }
     }
 
+    /// Sets the auto-moderation activation status
     pub async fn set_is_activate_auto_moderate(&self, data: bool) -> domain::services::ServiceResult<()> {
         self.repo.set("is_activate_auto_moderate", data).await
     }
